changed amount per day in habit

+++ lib/Model/Habit.dart
@@ -2,7 +2,7 @@ class Habit{
   String name;
   String description;
   DateTime startDate;
-  int amountPerDay;
+  double amountPerDay;
 
   Habit(this.name, this.description, this.startDate, this.amountPerDay);
 }
